replace ses-types-test with a ses/index.test-d.ts
