--- packages/ses-types-test/CHANGELOG.md
@@ -1,149 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-### [0.1.18](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.17...@endo/ses-types-test@0.1.18) (2022-03-07)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.17](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.16...@endo/ses-types-test@0.1.17) (2022-03-02)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.16](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.15...@endo/ses-types-test@0.1.16) (2022-02-20)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.15](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.14...@endo/ses-types-test@0.1.15) (2022-02-18)
-
-
-### Bug Fixes
-
-* Make sure lint:type runs correctly in CI ([a520419](https://github.com/endojs/endo/commit/a52041931e72cb7b7e3e21dde39c099cc9f262b0))
-* Unify TS version to ~4.2 ([5fb173c](https://github.com/endojs/endo/commit/5fb173c05c9427dca5adfe66298c004780e8b86c))
-
-
-
-### [0.1.14](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.13...@endo/ses-types-test@0.1.14) (2022-01-31)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.13](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.12...@endo/ses-types-test@0.1.13) (2022-01-27)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.12](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.11...@endo/ses-types-test@0.1.12) (2022-01-25)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.11](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.10...@endo/ses-types-test@0.1.11) (2022-01-23)
-
-
-### Bug Fixes
-
-* **ses:** Add assert.error options bag to type definition ([#978](https://github.com/endojs/endo/issues/978)) ([ca42997](https://github.com/endojs/endo/commit/ca4299714d5769ea15418612f679abb400ff7e25)), closes [#977](https://github.com/endojs/endo/issues/977)
-
-
-
-### [0.1.10](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.9...@endo/ses-types-test@0.1.10) (2021-12-14)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.9](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.8...@endo/ses-types-test@0.1.9) (2021-12-08)
-
-
-### Bug Fixes
-
-* Avoid eslint globs for Windows ([4b4f3cc](https://github.com/endojs/endo/commit/4b4f3ccaf3f5e8d53faefb4264db343dd603bf80))
-
-
-
-### [0.1.8](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.7...@endo/ses-types-test@0.1.8) (2021-11-16)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.7](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.6...@endo/ses-types-test@0.1.7) (2021-11-02)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.6](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.5...@endo/ses-types-test@0.1.6) (2021-10-15)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.5](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.4...@endo/ses-types-test@0.1.5) (2021-09-18)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.4](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.3...@endo/ses-types-test@0.1.4) (2021-08-14)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.3](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.2...@endo/ses-types-test@0.1.3) (2021-08-13)
-
-**Note:** Version bump only for package @endo/ses-types-test
-
-
-
-
-
-### [0.1.2](https://github.com/endojs/endo/compare/@endo/ses-types-test@0.1.1...@endo/ses-types-test@0.1.2) (2021-07-22)
-
-
-### Bug Fixes
-
-* **ses:** Fix assert type assertions ([53d284d](https://github.com/endojs/endo/commit/53d284d04eebed57ccaf19b43a1ff9a71393cc6b))
-
-
-
-### 0.1.1 (2021-06-20)
-
-**Note:** Version bump only for package @endo/ses-types-test
