--- packages/ses-types-test/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@endo/ses-types-test",
-  "version": "0.1.18",
-  "private": true,
-  "description": "TypeScript validation for SES types.",
-  "keywords": [],
-  "author": "Endo contributors",
-  "license": "Apache-2.0",
-  "homepage": "https://github.com/endojs/endo/tree/master/packages/ses-types-test#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/endojs/endo.git"
-  },
-  "bugs": {
-    "url": "https://github.com/endojs/endo/issues"
-  },
-  "type": "module",
-  "exports": {},
-  "scripts": {
-    "build": "exit 0",
-    "lint": "yarn lint:types && yarn lint:ts",
-    "lint-fix": "eslint --fix . --ext .ts",
-    "lint:ts": "eslint . --ext .ts",
-    "lint:types": "tsc -p tsconfig.json",
-    "test": "yarn lint:types"
-  },
-  "dependencies": {
-    "ses": "^0.15.11"
-  },
-  "devDependencies": {
-    "@endo/eslint-config": "^0.4.6",
-    "babel-eslint": "^10.0.3",
-    "eslint": "^7.23.0",
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-config-prettier": "^6.9.0",
-    "eslint-plugin-eslint-comments": "^3.1.2",
-    "eslint-plugin-import": "^2.19.1",
-    "eslint-plugin-prettier": "^3.1.2",
-    "prettier": "^1.19.1",
-    "typescript": "~4.6.2"
-  },
-  "files": [],
-  "publishConfig": {
-    "access": "private"
-  },
-  "eslintConfig": {
-    "extends": [
-      "@endo"
-    ]
-  },
-  "prettier": {
-    "trailingComma": "all",
-    "singleQuote": true
-  }
-}
