--- packages/ses-types-test/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "noEmit": true,
-    "target": "esnext",
-    "module": "esnext",
-    "moduleResolution": "node",
-    "strict": true,
-    "strictNullChecks": true,
-    "noImplicitAny": true,
-    "downlevelIteration": true,
-    "strictNullChecks": true
-  },
-  "include": ["**/*.ts", "**/*.js"]
-}
